Extract memory methods from Module.cpp
