+++ csrc/npu/NPUFunctions.cpp
@@ -187,4 +187,9 @@ aclrtContext GetDeviceContext(c10::DeviceIndex device) {
 aclError ResetUsedDevices() {
   return acl_adapter::ResetUsedDevices();
 }
+
+std::mutex* getFreeMutex() {
+  static std::mutex npu_free_mutex;
+  return &npu_free_mutex;
+}
 } // namespace c10_npu
