+++ torch_npu/csrc/Module.cpp
@@ -6,10 +6,11 @@
 #include "csrc/npu/NPUCachingHostAllocator.h"
 #include "npu/core/npu_log.h"
 #include "npu/core/sys_ctrl/npu_sys_ctrl.h"
-#include "torch_npu/csrc/npu/Module.h"
-#include "torch_npu/csrc/npu/Device.h"
 #include "torch_npu/csrc/core/AutocastMode.h"
 #include "torch_npu/csrc/core/TensorType.h"
+#include "torch_npu/csrc/npu/Device.h"
+#include "torch_npu/csrc/npu/Memory.h"
+#include "torch_npu/csrc/npu/Module.h"
 
 PyObject* module;
 
@@ -92,6 +93,7 @@ PyObject* initModule() {
 
   AddPyMethodDefs(methods, TorchNpuMethods);
   AddPyMethodDefs(methods, THNPModule_device_methods());
+  AddPyMethodDefs(methods, THNPModule_memory_methods());
   AddPyMethodDefs(methods, THNPModule_get_methods());
   AddPyMethodDefs(methods, torch_npu::utils::npu_extension_functions());
   AddPyMethodDefs(methods, torch_npu::autocast::autocast_mode_functions());
