+++ torch_npu/csrc/npu/Memory.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <Python.h>
+#include "csrc/core/Macros.h"
+
+TORCH_BACKEND_API PyMethodDef* THNPModule_memory_methods();
